requirements.txt
requests==2.32.3
python-dotenv==1.1.0
pandas==2.2.1
numpy==1.26.4
rich==13.7.0
eth-account==0.13.7
eth-utils==5.3.0
cryptography==42.0.5 